requirements.txt
numpy==1.9.0
scipy==0.13.3
bottle
matplotlib
XlsxWriter
openpyxl
sympy
Pillow